refactor: switched to u8

+++ scripts/importer/src/hook.rs
@@ -2,7 +2,7 @@ use serde::Serialize;
 
 #[derive(Serialize)]
 pub struct Hook {
-    pub id: u16,
+    pub id: u8,
     pub title: String,
     pub description: String,
     pub content: String,
